Move CGRect extension into separate file
